refactor: change path of di to sample01 (#1110)
