+++ advanced/di/sample01/di.go
@@ -1,4 +1,4 @@
-package di
+package sample01
 
 func Initialize() {
 	e := InitializeEvent("Hi there!")
