+++ advanced/di/sample01/di_test.go
@@ -1,4 +1,4 @@
-package di
+package sample01
 
 import (
 	"testing"
